requirements.txt
markdown 
python-docx 
beautifulsoup4
